+++ executorlib/backend/cache_parallel.py
@@ -4,7 +4,7 @@ import time
 
 import cloudpickle
 
-from executorlib.cache.backend import backend_load_file, backend_write_file
+from executorlib.task_scheduler.file import backend_load_file, backend_write_file
 
 
 def main() -> None:
