+++ executorlib/executor/flux.py
@@ -1,10 +1,10 @@
 import contextlib
 from typing import Callable, Optional, Union
 
-from executorlib.base.executor import ExecutorBase
-from executorlib.interactive.blockallocation import BlockAllocationTaskScheduler
-from executorlib.interactive.dependency import DependencyTaskScheduler
-from executorlib.interactive.onetoone import OneProcessTaskScheduler
+from executorlib.executor.base import ExecutorBase
+from executorlib.task_scheduler.interactive.blockallocation import BlockAllocationTaskScheduler
+from executorlib.task_scheduler.interactive.dependency import DependencyTaskScheduler
+from executorlib.task_scheduler.interactive.onetoone import OneProcessTaskScheduler
 from executorlib.standalone.inputcheck import (
     check_command_line_argument_lst,
     check_init_function,
@@ -16,7 +16,7 @@ from executorlib.standalone.inputcheck import (
 )
 
 with contextlib.suppress(ImportError):
-    from executorlib.interactive.fluxspawner import (
+    from executorlib.task_scheduler.interactive.fluxspawner import (
         FluxPythonSpawner,
         validate_max_workers,
     )
@@ -34,7 +34,7 @@ class FluxJobExecutor(ExecutorBase):
         max_workers (int): for backwards compatibility with the standard library, max_workers also defines the number of
                            cores which can be used in parallel - just like the max_cores parameter. Using max_cores is
                            recommended, as computers have a limited number of compute cores.
-        cache_directory (str, optional): The directory to store cache files. Defaults to "cache".
+        cache_directory (str, optional): The directory to store file files. Defaults to "file".
         max_cores (int): defines the number cores which can be used in parallel
         resource_dict (dict): A dictionary of resources required by the task. With the following keys:
                               - cores (int): number of MPI cores to be used for each function call
@@ -70,7 +70,7 @@ class FluxJobExecutor(ExecutorBase):
     Examples:
         ```
         >>> import numpy as np
-        >>> from executorlib.interfaces.flux import FluxJobExecutor
+        >>> from executorlib.executor.flux import FluxJobExecutor
         >>>
         >>> def calc(i, j, k):
         >>>     from mpi4py import MPI
@@ -118,7 +118,7 @@ class FluxJobExecutor(ExecutorBase):
             max_workers (int): for backwards compatibility with the standard library, max_workers also defines the
                                number of cores which can be used in parallel - just like the max_cores parameter. Using
                                max_cores is recommended, as computers have a limited number of compute cores.
-            cache_directory (str, optional): The directory to store cache files. Defaults to "cache".
+            cache_directory (str, optional): The directory to store file files. Defaults to "file".
             max_cores (int): defines the number cores which can be used in parallel
             resource_dict (dict): A dictionary of resources required by the task. With the following keys:
                                   - cores (int): number of MPI cores to be used for each function call
@@ -219,7 +219,7 @@ class FluxClusterExecutor(ExecutorBase):
         max_workers (int): for backwards compatibility with the standard library, max_workers also defines the number of
                            cores which can be used in parallel - just like the max_cores parameter. Using max_cores is
                            recommended, as computers have a limited number of compute cores.
-        cache_directory (str, optional): The directory to store cache files. Defaults to "cache".
+        cache_directory (str, optional): The directory to store file files. Defaults to "file".
         max_cores (int): defines the number cores which can be used in parallel
         resource_dict (dict): A dictionary of resources required by the task. With the following keys:
                               - cores (int): number of MPI cores to be used for each function call
@@ -251,7 +251,7 @@ class FluxClusterExecutor(ExecutorBase):
     Examples:
         ```
         >>> import numpy as np
-        >>> from executorlib.interfaces.flux import FluxClusterExecutor
+        >>> from executorlib.executor.flux import FluxClusterExecutor
         >>>
         >>> def calc(i, j, k):
         >>>     from mpi4py import MPI
@@ -296,7 +296,7 @@ class FluxClusterExecutor(ExecutorBase):
             max_workers (int): for backwards compatibility with the standard library, max_workers also defines the
                                number of cores which can be used in parallel - just like the max_cores parameter. Using
                                max_cores is recommended, as computers have a limited number of compute cores.
-            cache_directory (str, optional): The directory to store cache files. Defaults to "cache".
+            cache_directory (str, optional): The directory to store file files. Defaults to "file".
             max_cores (int): defines the number cores which can be used in parallel
             resource_dict (dict): A dictionary of resources required by the task. With the following keys:
                                   - cores (int): number of MPI cores to be used for each function call
@@ -341,7 +341,7 @@ class FluxClusterExecutor(ExecutorBase):
             {k: v for k, v in default_resource_dict.items() if k not in resource_dict}
         )
         if not plot_dependency_graph:
-            from executorlib.cache.task_scheduler import create_file_executor
+            from executorlib.task_scheduler.file import create_file_executor
 
             super().__init__(
                 executor=create_file_executor(
@@ -406,7 +406,7 @@ def create_flux_executor(
                            number of cores which can be used in parallel - just like the max_cores parameter. Using
                            max_cores is recommended, as computers have a limited number of compute cores.
         max_cores (int): defines the number cores which can be used in parallel
-        cache_directory (str, optional): The directory to store cache files. Defaults to "cache".
+        cache_directory (str, optional): The directory to store file files. Defaults to "file".
         resource_dict (dict): A dictionary of resources required by the task. With the following keys:
                               - cores (int): number of MPI cores to be used for each function call
                               - threads_per_core (int): number of OpenMP threads to be used for each function call
