+++ executorlib/standalone/hdf.py
@@ -108,7 +108,7 @@ def get_queue_id(file_name: Optional[str]) -> Optional[int]:
 def get_cache_data(cache_directory: str) -> list[dict]:
     file_lst = []
     for task_key in os.listdir(cache_directory):
-        file_name = os.path.join(cache_directory, task_key, "cache.h5out")
+        file_name = os.path.join(cache_directory, task_key, "file.h5out")
         os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
         if os.path.exists(file_name):
             with h5py.File(file_name, "r") as hdf:
