+++ executorlib/task_scheduler/file/backend.py
@@ -2,7 +2,7 @@ import os
 import time
 from typing import Any
 
-from executorlib.cache.shared import FutureItem
+from executorlib.task_scheduler.file.shared import FutureItem
 from executorlib.standalone.hdf import dump, load
