+++ executorlib/task_scheduler/interactive/onetoone.py
@@ -2,14 +2,14 @@ import queue
 from threading import Thread
 from typing import Optional
 
-from executorlib.base.task_scheduler import TaskSchedulerBase
-from executorlib.interactive.shared import execute_tasks
+from executorlib.task_scheduler.base import TaskSchedulerBase
+from executorlib.task_scheduler.interactive.shared import execute_tasks
 from executorlib.standalone.interactive.spawner import BaseSpawner, MpiExecSpawner
 
 
 class OneProcessTaskScheduler(TaskSchedulerBase):
     """
-    The executorlib.interactive.executor.InteractiveStepExecutor leverages the executorlib interfaces to distribute python
+    The executorlib.interactive.executor.InteractiveStepExecutor leverages the executorlib executor to distribute python
     tasks. In contrast to the mpi4py.futures.MPIPoolExecutor the executorlib.interactive.executor.InteractiveStepExecutor
     can be executed in a serial python process and does not require the python script to be executed with MPI.
     Consequently, it is primarily an abstraction of its functionality to improve the usability in particular when used
