+++ tests/test_cache_backend_execute.py
@@ -5,8 +5,8 @@ import unittest
 
 
 try:
-    from executorlib.cache.backend import backend_execute_task_in_file
-    from executorlib.cache.shared import _check_task_output, FutureItem
+    from executorlib.task_scheduler.file import backend_execute_task_in_file
+    from executorlib.task_scheduler.file import _check_task_output, FutureItem
     from executorlib.standalone.hdf import dump, get_runtime
     from executorlib.standalone.serialize import serialize_funct_h5
 
@@ -28,14 +28,14 @@ def get_error(a):
 )
 class TestSharedFunctions(unittest.TestCase):
     def test_execute_function_mixed(self):
-        cache_directory = os.path.abspath("cache")
+        cache_directory = os.path.abspath("file")
         os.makedirs(cache_directory, exist_ok=True)
         task_key, data_dict = serialize_funct_h5(
             fn=my_funct,
             fn_args=[1],
             fn_kwargs={"b": 2},
         )
-        file_name = os.path.join(cache_directory, task_key, "cache.h5in")
+        file_name = os.path.join(cache_directory, task_key, "file.h5in")
         os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
         dump(file_name=file_name, data_dict=data_dict)
         backend_execute_task_in_file(file_name=file_name)
@@ -46,24 +46,24 @@ class TestSharedFunctions(unittest.TestCase):
         self.assertTrue(future_obj.done())
         self.assertEqual(future_obj.result(), 3)
         self.assertTrue(
-            get_runtime(file_name=os.path.join(cache_directory, task_key, "cache.h5out"))
+            get_runtime(file_name=os.path.join(cache_directory, task_key, "file.h5out"))
             > 0.0
         )
         future_file_obj = FutureItem(
-            file_name=os.path.join(cache_directory, task_key, "cache.h5out")
+            file_name=os.path.join(cache_directory, task_key, "file.h5out")
         )
         self.assertTrue(future_file_obj.done())
         self.assertEqual(future_file_obj.result(), 3)
 
     def test_execute_function_args(self):
-        cache_directory = os.path.abspath("cache")
+        cache_directory = os.path.abspath("file")
         os.makedirs(cache_directory, exist_ok=True)
         task_key, data_dict = serialize_funct_h5(
             fn=my_funct,
             fn_args=[1, 2],
             fn_kwargs={},
         )
-        file_name = os.path.join(cache_directory, task_key, "cache.h5in")
+        file_name = os.path.join(cache_directory, task_key, "file.h5in")
         os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
         dump(file_name=file_name, data_dict=data_dict)
         backend_execute_task_in_file(file_name=file_name)
@@ -74,24 +74,24 @@ class TestSharedFunctions(unittest.TestCase):
         self.assertTrue(future_obj.done())
         self.assertEqual(future_obj.result(), 3)
         self.assertTrue(
-            get_runtime(file_name=os.path.join(cache_directory, task_key, "cache.h5out"))
+            get_runtime(file_name=os.path.join(cache_directory, task_key, "file.h5out"))
             > 0.0
         )
         future_file_obj = FutureItem(
-            file_name=os.path.join(cache_directory, task_key, "cache.h5out")
+            file_name=os.path.join(cache_directory, task_key, "file.h5out")
         )
         self.assertTrue(future_file_obj.done())
         self.assertEqual(future_file_obj.result(), 3)
 
     def test_execute_function_kwargs(self):
-        cache_directory = os.path.abspath("cache")
+        cache_directory = os.path.abspath("file")
         os.makedirs(cache_directory, exist_ok=True)
         task_key, data_dict = serialize_funct_h5(
             fn=my_funct,
             fn_args=[],
             fn_kwargs={"a": 1, "b": 2},
         )
-        file_name = os.path.join(cache_directory, task_key, "cache.h5in")
+        file_name = os.path.join(cache_directory, task_key, "file.h5in")
         os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
         dump(file_name=file_name, data_dict=data_dict)
         backend_execute_task_in_file(file_name=file_name)
@@ -102,24 +102,24 @@ class TestSharedFunctions(unittest.TestCase):
         self.assertTrue(future_obj.done())
         self.assertEqual(future_obj.result(), 3)
         self.assertTrue(
-            get_runtime(file_name=os.path.join(cache_directory, task_key, "cache.h5out"))
+            get_runtime(file_name=os.path.join(cache_directory, task_key, "file.h5out"))
             > 0.0
         )
         future_file_obj = FutureItem(
-            file_name=os.path.join(cache_directory, task_key, "cache.h5out")
+            file_name=os.path.join(cache_directory, task_key, "file.h5out")
         )
         self.assertTrue(future_file_obj.done())
         self.assertEqual(future_file_obj.result(), 3)
 
     def test_execute_function_error(self):
-        cache_directory = os.path.abspath("cache")
+        cache_directory = os.path.abspath("file")
         os.makedirs(cache_directory, exist_ok=True)
         task_key, data_dict = serialize_funct_h5(
             fn=get_error,
             fn_args=[],
             fn_kwargs={"a": 1},
         )
-        file_name = os.path.join(cache_directory, task_key, "cache.h5in")
+        file_name = os.path.join(cache_directory, task_key, "file.h5in")
         os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
         dump(file_name=file_name, data_dict=data_dict)
         backend_execute_task_in_file(file_name=file_name)
@@ -131,16 +131,16 @@ class TestSharedFunctions(unittest.TestCase):
         with self.assertRaises(ValueError):
             future_obj.result()
         self.assertTrue(
-            get_runtime(file_name=os.path.join(cache_directory, task_key, "cache.h5out"))
+            get_runtime(file_name=os.path.join(cache_directory, task_key, "file.h5out"))
             > 0.0
         )
         future_file_obj = FutureItem(
-            file_name=os.path.join(cache_directory, task_key, "cache.h5out")
+            file_name=os.path.join(cache_directory, task_key, "file.h5out")
         )
         self.assertTrue(future_file_obj.done())
         with self.assertRaises(ValueError):
             future_file_obj.result()
 
     def tearDown(self):
-        if os.path.exists("cache"):
-            shutil.rmtree("cache")
+        if os.path.exists("file"):
+            shutil.rmtree("file")
