+++ tests/test_fluxjobexecutor_plot.py
@@ -10,7 +10,7 @@ from executorlib.standalone.serialize import cloudpickle_register
 try:
     import pygraphviz
     import flux.job
-    from executorlib.interactive.fluxspawner import FluxPythonSpawner
+    from executorlib.task_scheduler.interactive import FluxPythonSpawner
 
     skip_graphviz_flux_test = "FLUX_URI" not in os.environ
 except ImportError:
