+++ tests/test_singlenodeexecutor_dependencies.py
@@ -5,8 +5,8 @@ from queue import Queue
 from threading import Thread
 
 from executorlib import SingleNodeExecutor
-from executorlib.interfaces.single import create_single_node_executor
-from executorlib.interactive.dependency import _execute_tasks_with_dependencies
+from executorlib.executor.single import create_single_node_executor
+from executorlib.task_scheduler.interactive.dependency import _execute_tasks_with_dependencies
 from executorlib.standalone.serialize import cloudpickle_register
 from executorlib.standalone.interactive.spawner import MpiExecSpawner
